backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
httpx>=0.27.0
aiosqlite>=0.20.0
sse-starlette>=2.0.0
chromadb>=0.5.0
openai>=1.0.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
